+++ src/aks-preview/HISTORY.rst
@@ -12,6 +12,10 @@ To release a new version, please select a new version number (usually plus 1 to 
 Pending
 +++++++
 
+18.0.0b22
++++++++
+* Vendor new SDK and bump API version to 2025-06-02-preview.
+
 18.0.0b21
 +++++++
 * Add command `az aks bastion` to enable connections to managed Kubernetes clusters via Azure Bastion.